+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Unreleased
 
+- Suppress CROTA warnings in load_ndcube_from_fits in https://github.com/punch-mission/punchbowl/pull/456
 - Corrects the sign in p angle when converting between helio and celestial in https://github.com/punch-mission/punchbowl/pull/454
 
 ## Version 0.0.12: May 12, 2025